app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
python-jose
python-multipart